Move navbar from molecules to organisms
